[수정] HttpRequest 구조가 불필요하게 복잡하여 구조 수정
- HttpRequestImpl과 HttpRequestBuilderImpl이 불필요하다고 고려되어 삭제 후 HttpRequest class 수정
