+++ src/main/java/webserver/HttpRequestParser.java
@@ -26,6 +26,7 @@ public class HttpRequestParser {
     private static void parseHeaders(HttpRequest.Builder builder, BufferedReader bufferedReader) throws IOException {
         String oneLine = bufferedReader.readLine();
         while (oneLine.length() != 0) {
+            // TODO: Host에 : 들어오면 split 이상하게 나뉠 수 있음.
             String[] header = oneLine.split(":");
             builder.setHeader(header[0].trim(), header[1].trim());
 //            logger.info("HEADERS: {}", oneLine);
